+++ pyyorkshirewater/__init__.py
@@ -41,7 +41,7 @@ from .models import (
     YearlyConsumption,
 )
 
-__version__ = "1.3.0"
+__version__ = "1.4.0"
 
 __all__ = [
     "IDP_COOKIE_DOMAIN",